:memo: Add docs for dependencies with yield
